### 2024-11-18 #7496 @editorialbot
[PRE REVIEW]: NMRInversions.jl, a julia package for time-domain Nuclear Magnetic Resonance
**Submitting author:** @aris-mav (<a href="http://orcid.org/0000-0002-6619-2303">Aristarchos Mavridis</a>)
**Repository:** https://github.com/aris-mav/NMRInversions.jl
**Branch with paper.md** (empty if default branch): master
**Version:** v0.9.2
