### 2024-11-18 #7496 @kyleniemeyer
Hello @aris-mav thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.

While we are waiting, could you please:

- Fix the DOIs (noting that @editorialbot's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references:

- Recommend some potential reviewers from the database, though please do not tag them with @ directly.
