### 2024-11-19 #7496 @aris-mav
Hello Kyle,
Thank you very much for the information.

As for the reviewers, should I only look at the current list, or should I also ask people in the field who might be willing to review?
From the list, the only person into magnetic resonance seems to be [nicksisco1932](https://reviewers.joss.theoj.org/reviewers/1755)  (based on a quick keyword search).
