### 2024-11-19 6c0cc91
paper updates
