+++ paper/paper.bib
@@ -7,7 +7,7 @@
     pages={65--98},
     year={2017},
     publisher={SIAM},
-    url={https://doi.org/10.1137/141000671}
+    doi={10.1137/141000671}
 }
 
 @article{Fordham2017,
@@ -34,7 +34,7 @@
 } 
 
 
-@book{Hansen_2010, 
+@book{Hansen2010, 
     title={Discrete inverse problems: Insight and algorithms}, 
     publisher={Society for Industrial and Applied Mathematics}, 
     author={Hansen, Per Christian}, 
