### 2024-11-28 fd8c604
plot expfits changes
