+++ ext/gui_ext.jl
@@ -13,9 +13,11 @@ e.g. `plot(data1)` plots the data and the corresponding fit,
 but `plot(data1, data2)` or `plot(data1, data2, data3)` work as well,
 and they plot all of the results on the same plot.
 
-This function creates a figure, 
-and calls `plot!(fig, res)` for each of the results in `res`,
-so please refer to the `plot!` documentation for more information.
+Keyword (optional) arguments:
+- `markersize` : The size of the markers (default is 7).
+- `normeq` : Whether to plot the normalised form of the equation or not (default is `true`).
+- `yscale` : The scale of the y-axis (default is `identity`).
+
 """
 function Makie.plot(res::expfit_struct...; kwargs...)
 
@@ -24,6 +26,16 @@ function Makie.plot(res::expfit_struct...; kwargs...)
 
     return f
 
+end
+
+function Makie.plot(res_mat::AbstractVecOrMat{NMRInversions.expfit_struct}; kwargs...)
+
+    f = Figure(size=(800, 400))
+    plot!(f, res_mat...; kwargs...)
+
+    return f
+
+end
 end
 
 """
@@ -39,6 +51,7 @@ The arguments are:
 Keyword (optional) arguments:
 - `markersize` : The size of the markers (default is 7).
 - `normeq` : Whether to plot the normalised form of the equation or not (default is `true`).
+- `yscale` : The scale of the y-axis (default is `identity`).
 
 Note that the res inputs are not a vector, but individual `expfit_struct` structures,
 like: plot!(fig, data1, data2, data3).
@@ -46,17 +59,17 @@ If you want to use a vector of `expfit_struct` structures, make sure to
 splat it by using `...` in the function call (e.g. `plot!(fig, [data1, data2, data3]...)`).
 """
 function Makie.plot!(fig::Union{Makie.Figure,Makie.GridPosition}, res::expfit_struct...;
-     markersize=7, normeq=true)
+     markersize=7, normeq=true, yscale = identity)
 
     # Make axes
     if res[1].seq in [NMRInversions.IR]
-        ax = Axis(fig[1, 1], xlabel="time (s)", ylabel="Signal (a.u.)")
+        ax = Axis(fig[1, 1], xlabel="time (s)", ylabel="Signal (a.u.)",yscale=yscale)
 
     elseif res[1].seq in [NMRInversions.CPMG]
-        ax = Axis(fig[1, 1], xlabel="time (s)", ylabel="Signal (a.u.)")
+        ax = Axis(fig[1, 1], xlabel="time (s)", ylabel="Signal (a.u.)",yscale=yscale)
 
     elseif res[1].seq in [NMRInversions.PFG]
-        ax = Axis(fig[1, 1], xlabel="b factor (s/m² e-9)", ylabel="Signal (a.u.)")
+        ax = Axis(fig[1, 1], xlabel="b factor (s/m² e-9)", ylabel="Signal (a.u.)",yscale=yscale)
     end
 
     for (i, r) in enumerate(res)
