+++ ext/gui_ext.jl
@@ -36,7 +36,6 @@ function Makie.plot(res_mat::AbstractVecOrMat{NMRInversions.expfit_struct}; kwar
     return f
 
 end
-end
 
 """
     plot!(fig, res...; names, markersize, normeq)
