### 2024-12-06 b4b1327
kernel svd bugfix for overdetermined problems
