+++ src/kernels.jl
@@ -33,10 +33,15 @@ to remove the "noisy" singular values.
 """
 function create_kernel(seq::Type{<:pulse_sequence1D}, x::Vector, X::Vector, g::Vector{<:Real})
 
-    usv = svd(create_kernel(seq, x , X))
-
-    K_new = Diagonal(usv.S) * usv.V'
-    g_new = usv.U' * g
+    if x < X
+        usv = svd(create_kernel(seq, x , X))
+        K_new = Diagonal(usv.S) * usv.V'
+        g_new = usv.U' * g
+    else
+        usv = svd(create_kernel(seq, x , X),full=true)
+        K_new = [Diagonal(usv.S) ; zeros(size(usv.U,1)-length(usv.S),length(usv.S))] * usv.V'
+        g_new = usv.U' * g
+    end
 
     return svd_kernel_struct(K_new, g_new, usv.U, usv.S, usv.V)
 
