### 2024-12-09 f51a6c6
added different norm options for expfits
