+++ src/exp_fits.jl
@@ -42,7 +42,7 @@ function mexp_loss(u, p)
     x = p[1]
     y = p[2]
     seq = p[3]
-    return sum((mexp(seq, u, x) .- y) .^ 2)
+    return norm((mexp(seq, u, x) .- y) , p[4])
 end
 
 
@@ -63,6 +63,7 @@ Arguments:
 Optional arguments:
 - `solver` : OptimizationOptimJL solver, defeault choice is BFGS().
 - `normalize` : Normalize the data before fitting? (default is true).
+- `L` : An integer specifying which norm of the residuals you want to minimize (default is 2).
 
 
 The `n` argument can also be a vector of initial parameter guesses, 
@@ -91,7 +92,8 @@ function expfit(
     x::Vector, 
     y::Vector;
     solver=OptimizationOptimJL.BFGS(),
-    normalize=true
+    normalize::Bool=true,
+    L::Int = 2
 )
 
     if normalize
@@ -116,7 +118,7 @@ function expfit(
 
     # Solve the optimization
     optf = Optimization.OptimizationFunction(mexp_loss, Optimization.AutoForwardDiff())
-    prob = Optimization.OptimizationProblem(optf, u0, (x, y, seq), lb=zeros(length(u0)), ub=Inf .* ones(length(u0)))
+    prob = Optimization.OptimizationProblem(optf, u0, (x, y, seq, L), lb=zeros(length(u0)), ub=Inf .* ones(length(u0)))
     u = OptimizationOptimJL.solve(prob, solver, maxiters=5000, maxtime=100)
 
     # Determine what's the x-axis of the seq (time or bfactor)
