### 2024-12-10 afb0452
improved xfit, yfit for invert and expfit
