+++ src/types.jl
@@ -283,6 +283,8 @@ The fields are as follows:
 - `seq`: The pulse sequence
 - `x` : The x acquisition values (e.g. time for relaxation or b-factor for diffusion).
 - `y` : The y acquisition values.
+- `xfit` : An array with fitted x values (for plotting purposes).
+- `yfit` : An array with fitted y values (for plotting purposes).
 - `u` : The fitted parameters for the `mexp` function.
 - `u0` : The initial parameters for the `mexp` function.
 - `r` : The residuals.
@@ -295,6 +297,8 @@ mutable struct expfit_struct
     seq::Type{<:NMRInversions.pulse_sequence1D}
     x::Vector
     y::Vector
+    xfit::Vector
+    yfit::Vector
     u::Vector
     u0::Vector
     r::Vector
