+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"functions/#Intro","page":"Functions","title":"Intro","text":"","category":"section"},{"location":"functions/","page":"Functions","title":"Functions","text":"This page contains the documentation for various useful  functions in the NMRInversions package.","category":"page"},{"location":"functions/","page":"Functions","title":"Functions","text":"tip: Tip\nFrom the julia command line, you can enter '?',  followed by the name of any function, struct,  or object you want to learn more about (try it!). After typing using NMRInversions in the julia console,  this feature will work for all the functions mentioned below.","category":"page"},{"location":"functions/","page":"Functions","title":"Functions","text":"info: Info\nIn julia, function definitions look like this:foo(x, y, z ; a, b)For the example above, foo is the name of the function,  and the contents of the parenteses are the arguments.   Within the parenteses, we got two types of arguments:Positional argumentsEverything that appears before the semicolon ;  (x,y and z in this example) is necessary, and must be given in a specific order.Keyword arguments.Everything that appears after the semicolon ;  (b and a in this example) is optional, and can be given in any order, but its name must be specified.In the example above, we can call the function by typing foo(1, 2, 3), (in which case, x=1, y=2, z=3, and the default value for a and b will be used).  You can also call the function by typing foo(1, 2, 3, a=3), to specify the value for a,  or by typing foo(1, 2, 3, b=3, a = 2), to specify the value for both a and b.Sometimes, if there are many keyword arguments, we write  the function as foo(x, y ; kwargs...).  For the necessary arguments before the semicolon, order matters.  For the keyword arguments after the semicolon, order does not matter,  but the name of each argument must be specified. For more information, please refer to this link.","category":"page"},{"location":"functions/#Importing-data-functions","page":"Functions","title":"Importing data functions","text":"","category":"section"},{"location":"functions/","page":"Functions","title":"Functions","text":"This package offers some functions to import NMR experiment data of various formats. Alternatively, you can of course import your data however you see fit. If a format you're working with is not yet supported,  please submit an issue  and we'll work on it.","category":"page"},{"location":"functions/","page":"Functions","title":"Functions","text":"The most basic use case would be using data saved in a csv format,  where there are only two columns,  one for your x-axis (time for relaxation and b-factor for diffusion) and one for your y-axis (signal intensity).","category":"page"},{"location":"functions/","page":"Functions","title":"Functions","text":"import_csv(::Type{<:pulse_sequence1D},::String)","category":"page"},{"location":"functions/#NMRInversions.import_csv-Tuple{Type{<:pulse_sequence1D}, String}","page":"Functions","title":"NMRInversions.import_csv","text":"import_csv(seq, file)\n\nImport data from a CSV file. The function reads the file and returns an input1D structure.\n\nseq is the 1D pulse sequence (e.g. IR, CPMG, PFG)\nfile is the path to the CSV file which contains the data (x, y) in two respective columns.\n\nThe function can be called without the seq argument, and the output will be the x and y vectors  (use it like, x,y =import_csv()). Alternatively, the function can also be called with only the seq argument, in which case a file dialog will open to select the file (use it like, data = import_csv(IR))\n\nPlease note that this function will just import the data as is, without any unit conversions. Ensure tha
[…]
