### 2025-01-02 #7496 @zhubonan
Hi @luthaf and @nicksisco1932, and @nluetts - Happy New Year! Would you be happy to review this submission for JOSS?

JOSS is unique and takes place on GitHub. This involves reviewing the manuscript and installing/testing/reading the underlying code. You can learn more about it here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#reviewing-for-joss

Please let me know if you are able/unable to review. Your response is greatly appreciated. Thanks in advance.
