### 2025-01-03 984237e
swapped to optim, refactored find_alpha, tests pass
