+++ src/optim_regularizations.jl
@@ -1,14 +1,20 @@
+function solve_regularization(K::AbstractMatrix, g::AbstractVector, α::Real, solver::brd)
 
-function solve_regularization(K::AbstractMatrix, g::AbstractVector, α::Real, solver::Type{brd})
+    f = x -> objective_f(x, (α, g, K))
+    g! = (grad, x) -> gradient_f!(grad, x, (α, g, K))
+    h! = (hess, x) -> hessian_f!(hess, x, (α, g, K))
 
-    optf = Optimization.OptimizationFunction(objective_f, grad=gradient_f, hess=hessian_f)
-    prob = Optimization.OptimizationProblem(optf, ones(size(K, 1)), (α, g, K))
-    c = OptimizationOptimJL.solve(prob, OptimizationOptimJL.NewtonTrustRegion(), x_tol=1e-8, maxiters=5000, maxtime=100)
+    c = Optim.optimize(
+        f, g!, h!,
+        ones(size(K, 1)),
+        NewtonTrustRegion(),
+        Optim.Options(x_tol = 1e-8)
+    ).minimizer
+
     f = vec(max.(0, (K' * c)))
     r = K * f - g
 
     return f, r
-
 end
 
 function objective_f(c::AbstractVector, p=(α, data, K))
@@ -16,12 +22,14 @@ function objective_f(c::AbstractVector, p=(α, data, K))
     G[:, (p[3]'*c.<=0)] .= 0
     f = 0.5 * c' * ((p[1] * I + (G * p[3]')) * c) - c' * p[2]
 end
-function gradient_f(gradient, c, p=(α, data, K))
+
+function gradient_f!(gradient, c, p=(α, data, K))
     G = copy(p[3])
     G[:, (p[3]'*c.<=0)] .= 0
     gradient .= (p[1] * I + G * p[3]') * c - p[2]
 end
-function hessian_f(H, c, p=(α, data, K))
+
+function hessian_f!(H, c, p=(α, data, K))
     G = copy(p[3])
     G[:, (p[3]'*c.<=0)] .= 0
     H .= p[1] * I + G * p[3]'
@@ -46,14 +54,14 @@ Solve a least squares problem, with nonnegativity constraints.
 """
 function solve_nnls(A::AbstractMatrix, b::AbstractVector)
 
-    optf = Optimization.OptimizationFunction(obj_f, Optimization.AutoForwardDiff())
-    prob = Optimization.OptimizationProblem(optf, ones(size(A, 2)), (A, b), lb=zeros(size(A, 2)), ub=Inf * ones(size(A, 2)))
-    x = OptimizationOptimJL.solve(prob, OptimizationOptimJL.LBFGS(), maxiters=5000, maxtime=100)
+    x = Optim.optimize(
+        x -> obj_ls(x, (A,b)), 
+        zeros(size(A, 2)), Inf .* ones(size(A, 2)), ones(size(A, 2)),
+        autodiff = :forward,
+        Optim.Options(x_tol = 1e-8)
+    ).minimizer
 
     return x
-end
-function obj_f(x, p)
-    return sum((p[1] * x - p[2]).^2)
 end
 
 
@@ -62,13 +70,16 @@ Solve a least squares problem, without nonnegativity constraints.
 """
 function solve_ls(A::AbstractMatrix, b::AbstractVector)
 
-    optf = Optimization.OptimizationFunction(obj_ls, Optimization.AutoForwardDiff())
-    prob = Optimization.OptimizationProblem(optf, ones(size(A, 2)), (A, b), 
-                                            lb= -Inf .* ones(size(A, 2)), ub=Inf .* ones(size(A, 2)))
[…]
