+++ src/NMRInversions.jl
@@ -6,12 +6,12 @@ using LinearAlgebra
 using SparseArrays
 using NativeFileDialog
 using PolygonOps
-import Optimization, OptimizationOptimJL
+using Optim 
 
 """
 to do list:
 - add gcv for reci method
-- differentiate between Mitchell GCV and optimization GCV
+- differentiate between Mitchell GCV and optim GCV
 - introduce faf, flip angle fraction, to the kernel functions. 1 would be a perfect pulse, 0 would be no pulse.
 - add precompilation
 

[…]
