### 2025-01-03 a4e7aed
added some tests
