### 2025-01-06 15e3304
documented find alpha constructors
