+++ docs/src/functions.md
@@ -118,6 +118,8 @@ lcurve(::Real, ::Real ; kwargs...)
 The `gcv()` method (from Mithcell 2012) usually involves the least amount of 
 function calls and it is thus much faster, thus used as the default option.
 If you want more precision, the univariate or box methods should be used instead.
+Note that `gcv()` will NOT work for `pdhgm()` solver, so you'll have to choose 
+an alternative explicitly when using that solver.
 
 # Exponential fit functions
 
