### 2025-01-22 6360855
normalization bugfix
