### 2025-01-22 cd32c4e
bump verson to 0.9.4
