### 2025-02-04 #7745 @editorialbot
[REVIEW]: NMRInversions.jl, a julia package for time-domain Nuclear Magnetic Resonance
**Submitting author:** @aris-mav (<a href="http://orcid.org/0000-0002-6619-2303">Aristarchos Mavridis</a>)
**Repository:** https://github.com/aris-mav/NMRInversions.jl
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.0
