### 2025-02-17 7ce350c
adding instructions for example data, closes #46
