### 2025-02-17 e8b5b50
added dt2 data and tutorial.jl, closes #47
