### 2025-02-21 cb7726c
added some types to the functions in solve_regularization
