### 2025-03-03 9286996
fixed typos and some formatting in documentation
