### 2025-03-03 2ed475e
temporary solution to mac window resize bug
