### 2025-03-06 66979d4
defined outline for delete_range function
