### 2025-03-06 3c609e8
added delete_range! to the exported functions
