+++ src/NMRInversions.jl
@@ -35,6 +35,7 @@ export import_spinsolve, import_geospec
 # The following functions are defined here and modified from extension files
 function select_peaks end
 function pubfig end
-export select_peaks, pubfig
+function delete_range! end
+export select_peaks, pubfig, delete_range!
 
 end
