### 2025-03-06 2c1be9c
delte_range! basic implementation
