+++ src/misc.jl
@@ -251,12 +251,15 @@ end
 
 """
     delete_range!(results::inv_res_1D , range)
-set selected range of f to 0, and update results
+Set selected range of f to 0, and update results.
 """
-function delete_range!(results::inv_res_1D , range)
+function delete_range!(results::inv_out_1D , range)
 
     # turn the section of `f` corresponding to the range to 0
     # new_g = K*f
     # results .= new_results
+
+    results.f[range[1]:range[2]] .= 0
+    # display(results.f)
 
 end
