### 2025-03-06 b7f3a53
added 'ignore selection' button to GUI
