### 2025-03-07 9893797
reset button resets ignored regions
