### 2025-03-07 89af5f7
update fitted curve when removing section of f
