### 2025-03-10 8b0dd07
delete selection button changes
