### 2025-03-11 5d00378
added ker_struct to inv_out_1D type
