+++ src/NMRInversions.jl
@@ -35,7 +35,7 @@ export import_spinsolve, import_geospec
 # The following functions are defined here and modified from extension files
 function select_peaks end
 function pubfig end
-function delete_range! end
-export select_peaks, pubfig, delete_range!
+function filter_selection! end
+export select_peaks, pubfig, filter_selection!
 
 end
