### 2025-03-18 d377ce6
updated filter parameter in 1D inversion results and residuals calculation
