### 2025-03-18 22c5e00
changing how residuals are calculated and plotte in 1D case
