### 2025-03-18 949ee0e
filter_selection only changes filter, changes to inv_out_1D
