### 2025-03-18 a127d43
fit and residuals updating moved from filter_selection! to just before plotting
