### 2025-03-19 a044e51
fixed PFG plot and filtered selections case
