+++ ext/gui_ext.jl
@@ -176,7 +176,7 @@ function Makie.plot(res::NMRInversions.inv_out_1D)
         empty!(fig.content[1])
         empty!(fig.content[2])
         empty!(fig.content[3])
-        filter_selection!(res, slider.interval[])
+        filter_range!(res, slider.interval[])
         draw_on_axes(fig.content[1], fig.content[2], fig.content[3], res, selections = true)
         vlines!(fig.content[2], int_low, color=:red)
         vlines!(fig.content[2], int_high, color=:red)
