### 2025-03-19 dbe46bb
'Remove Section' GUI implementation (#54)
* fixed typos and some formatting in documentation

* temporary solution to mac window resize bug

* defined outline for delete_range function

* added delete_range! to the exported functions

* delte_range! basic implementation

* added 'ignore selection' button to GUI

* reset button resets ignored regions

* bug fix

* update fitted curve when removing section of f

* delete_range! updates residuals

* delete selection button changes

* comments and docs change

* docs change

* removed temporary bug fix

* added ker_struct to inv_out_1D type

* f s
[…]
