+++ ext/Project.toml
@@ -1,0 +1,6 @@
+[deps]
+GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+NMRInversions = "55c20db2-0166-4687-95c3-62a9c7afb29b"
+NativeFileDialog = "e1fe445b-aa65-4df4-81c1-2041507f0fd4"
+PolygonOps = "647866c9-e3ac-4575-94e7-e3d426903924"
