+++ docs/make.jl
@@ -11,11 +11,12 @@ makedocs(sitename="NMRInversions.jl",
                  ],
 
          pages=["Overview" => "index.md",
+                "Theory" => "theory.md",
                 "Tutorial" => "tutorial.md",
                 "Functions" => "functions.md",
                 "Types and Structures" => "types_structs.md",
                 "Saving data" => "savefiles.md",
-                #="Theory" => "theory.md",=#
+                "References" => "references.md"
                ],
          checkdocs=:none
         )
