### 2025-03-22 8ba4815
add installation section
