+++ docs/src/types_structs.md
@@ -1,7 +1,7 @@
 
 # Types and Structures
 
-In order to take advantage of julia's multiple dispatch, 
+In order to take advantage of Julia's multiple dispatch, 
 we need to define some structures which can be used as input to our functions.
 
 !!! info
