### 2025-03-22 6e85776
added plots documentation
