+++ docs/src/index.md
@@ -2,7 +2,12 @@ Welcome to the NMRInversions.jl package!
 
 Hopefully this documentation will provide all the information you need to get started.
 
-    
-
-For more details on how to use the package, you can start by refering to the [tutorial](tutorial.md) section.
-
+- The [Theory](theory.md) section should provide some info on what the package is about.
+- For more details on how to use it, you can start by refering to the 
+  [tutorial](tutorial.md) section.
+- To understand how things really work, please have a look at the 
+  [Functions](functions.md) and [Types and Structures](types_structs.md) sections.
+- If you have any problems or questions, please feel free to 
+  [submit an issue](https://github.com/aris-mav/NMRInversions.jl/issues).
+- If you would like to contribute to the package, you may start with one of 
+  the [existing issues](https://github.com/aris-mav/NMRInversions.jl/issues).
