### 2025-03-23 #7745 @aris-mav
Thank you very much for the comments, @hessammehr !
I have created a [pull request](https://github.com/aris-mav/NMRInversions.jl/pull/57) to address these, so please have a look at that.

Indeed, I was not aware of the harminv library, thanks for bringing it up.  I'd be happy to mention it in the paper, but I can't find any cases where it's used for this purpose. Would there be any such examples I might be missing?
