### 2025-03-23 #7745 @hessammehr
That's great, I'll have a look at your PR shortly.

I would definitely encourage mentioning `harminv`. You are right that mentions of it in conjunction with NMR are sparse — I only found a couple of direct NMR applications on [Scholar](https://scholar.google.com/scholar?start=0&q=%22harminv%22+%22nmr%22) but still worth a mention.
