### 2025-03-23 #7745 @aris-mav
Thanks!
To clarify regarding `harminv`, I did find some general NMR uses, but they seem more geared towards spectroscopy-related applications rather than relaxation/diffusion. Just by looking at their documentation, I'm not entirely sure whether it handles cases where sinusoidal signals are not involved, such as the present one.
