### 2025-03-23 #7745 @aris-mav
> Regarding the docs, I would still advocate for a 1-2 sentence summary of what this package is right on the main docs page.

Would that refer to the documentation overview, or the README file?
