+++ docs/src/types_structs.md
@@ -28,7 +28,8 @@ IRCPMG
 ## Inversion solvers
 These are used to let the invert fucntion know which solver to use.
 They can be used as input to the `invert` function as the 'solver' argument.
-(e.g., `invert(data, solver=brd)` or `invert(data, solver=pdhgm(10,0.1) )`).
+(e.g., `invert(data, solver=brd)` [Song2002](@cite) or 
+`invert(data, solver=pdhgm(10,0.1) )`) [reci2017](@cite).
 
 ```@docs
 brd
