+++ docs/src/functions.md
@@ -115,7 +115,7 @@ lcurve(::Real; kwargs...)
 lcurve(::Real, ::Real ; kwargs...)
 ```
 
-The `gcv()` method (from Mithcell 2012) usually involves the least amount of 
+The `gcv()` method [Mitchell2012](@cite) usually involves the least amount of 
 function calls and it is thus much faster, thus used as the default option.
 If you want more precision, the univariate or box methods should be used instead.
 Note that `gcv()` will NOT work for `pdhgm()` solver, so you'll have to choose 
