+++ docs/src/index.md
@@ -1,11 +1,16 @@
 Welcome to the NMRInversions.jl package!
 
+Functionality includes :
+- Importing NMR data of various formats.
+- Performing numerical inversions with sensible default settings.
+- Interactive visualization through the Makie extension.
+
 Hopefully this documentation will provide all the information you need to get started.
 
 - The [Theory](theory.md) section should provide some info on what the package is about.
 - For more details on how to use it, you can start by refering to the 
   [tutorial](tutorial.md) section.
-- To understand how things really work, please have a look at the 
+- To understand how things really work, have a look at the 
   [Functions](functions.md) and [Types and Structures](types_structs.md) sections.
 - If you have any problems or questions, please feel free to 
   [submit an issue](https://github.com/aris-mav/NMRInversions.jl/issues).
