### 2025-04-08 #7745 @aris-mav
Hello all,

@hessammehr, would you please let me know whether any updates are needed in https://github.com/aris-mav/NMRInversions.jl/pull/57 before merging?

Also, could you please elaborate on the comment about the tests above?

@cfarm6, any comments? Please let me know if anything is unclear.
