### 2025-04-08 63c0e6c
added goals and how to contribute in overview
