### 2025-04-08 d307708
added why we need inversions
