### 2025-04-08 #7745 @aris-mav
@hessammehr Thanks for the prompt response! I just pushed a couple of commits to the [PR](https://github.com/aris-mav/NMRInversions.jl/pull/57) addressing remaining comments.

Regarding `harminv` , after going through its docs, I'm still not sure whether it can do the same job as this package.
Probably we refer to different "inversions" here.
It states :
```
In general, it is useful when you know on physical grounds that your system consists of a
small number of decaying & oscillating modes in the bandwidth of interest, plus a limited
amount of noise, and is not appropriate to analyze completely arbitrary waveforms.
```
I imagine even if we interpret a relaxation signal as "oscillating with zero frequency" and only want to uncover the decaying modes, harminv would still not cover the much needed regularization, which NMRInversions.jl handles by default.
Please let me know in case I'm missing anything here.
