+++ paper/paper.bib
@@ -53,3 +53,14 @@
   publisher = {American Physical Society},
   doi = {10.1103/PhysRev.73.679},
 }
+
+@article{mandel1997,
+    author = {Mandelshtam, Vladimir A. and Taylor, Howard S.},
+    title = {Harmonic inversion of time signals and its applications},
+    journal = {The Journal of Chemical Physics},
+    volume = {107},
+    number = {17},
+    pages = {6756-6769},
+    year = {1997},
+    month = {11},
+}
