+++ paper/paper.md
@@ -32,8 +32,13 @@ It is aimed towards NMR users who are not necessarily familiar with the Julia pr
 or even programming in general.
 
 Functionality includes importing data from various NMR instrument formats,
-performing phase-correction on raw data, fitting multi-exponential curves,
+performing phase-correction on raw data, fitting multi-exponential curves using numerical inversions,
 and interactively visualizing the results in a straightforward manner.
+
+It should be noted that the aforementioned inversions should not be confused with 
+harmonic inversions [@mandel1997] used in NMR spectroscopy. These can be handled by 
+other packages such as "harminv" (https://github.com/NanoComp/harminv), 
+and are not yet implemented in NMRInversions.jl at the time of writing.
 
 # Statement of need
 
