### 2025-04-08 #7745 @aris-mav
Thanks, just wanted to check whether we're on the same page as to whether harminv can achieve the same job or not.
I just pushed [some lines](https://github.com/aris-mav/NMRInversions.jl/pull/57/commits/819603a46d4ff01101f3171ee069b6e13c2c8376) on the paper to clarify this.
