### 2025-04-09 #7745 @hessammehr
Almost there — and sorry to be a bother — but I would advocate for a slightly expanded description in `index.md`. Who would need to use this package and for what applications/purposes/in what scenarios? It doesn't need to be exhaustive, of course, a few examples would suffice. I imagine people landing on the docs page for the first time would appreciate the extra context.

This is my only remaining suggestion.
