+++ docs/src/index.md
@@ -1,4 +1,4 @@
-Welcome to the NMRInversions.jl package!
+Welcome to NMRInversions.jl!
 
 The goal is to provide an easy-to-use interface to process NMR 
 relaxation and diffusion data, completely free/open-source, and fully 
@@ -9,11 +9,24 @@ Functionality includes :
 - Performing numerical inversions and exponential fits with sensible default settings.
 - Interactive visualization of results, through the GLMakie extension.
 
+Who would want to use this package :
+- Researchers in chemistry, physics, and materials science who rely on NMR 
+  to study molecular dynamics, diffusion and interactions. 
+  NMRInversions.jl simplifies the process of data analysis, allowing researchers 
+  to focus on their scientific questions rather than the intricacies of data processing.
+- Professionals in various fields such as food science, subsurface applications, 
+  or materials engineering can utilize this package to analyze NMR data for quality 
+  control, product development, and research purposes. 
+- Students learning about NMR techniques can benefit from a straightforward tool 
+  that helps them understand data processing and analysis. 
+  Educators can use the package as a teaching aid in courses related to NMR or 
+  inverse problems.
+
 Hopefully this documentation will provide all the information you need to get started.
 
-- The [Theory](theory.md) section should provide some info on what the package is about.
-- For more details on how to use it, you can start by refering to the 
-  [tutorial](tutorial.md) section.
+- The [Theory](theory.md) section should provide some info on what the package is really about.
+- For more details on how to use it, you can start by refering to the examples in the  
+  [tutorial](tutorial.md).
 - To understand how things really work, have a look at the 
   [Functions](functions.md) and [Types and Structures](types_structs.md) sections.
 - If you have any problems or questions, please feel free to 
