### 2025-04-09 #7745 @aris-mav
No bother at all! A third perspective is always useful. I've added [a few more lines](https://github.com/aris-mav/NMRInversions.jl/pull/57/commits/af1d58ab6881fe54df545410e06260afbe2e21aa) addressing this.
