### 2025-04-09 0fceeeb
JOSS review - Hessam comments (#57)
* typos addressed

* add installation section

* add installation section

* capitalize julia to Julia

* added plots documentation

* added joss status

* paper edits

* added expfit plot

* typo

* added references

* typo

* improved headings

* added goals and how to contribute in overview

* added why we need inversions

* added harminv clarification

* added target audience
