+++ paper/paper.md
@@ -1,11 +1,11 @@
 ---
-title: 'NMRInversions.jl, a julia package for time-domain Nuclear Magnetic Resonance'
+title: 'NMRInversions.jl, a Julia package for time-domain Nuclear Magnetic Resonance'
 date: 04 November 2024 
 tags:
   - Magnetic Resonance
   - NMR relaxation
   - NMR diffusion
-  - julia
+  - Julia
   - Inverse problems
   - Numerical inversion
 authors:
@@ -28,45 +28,51 @@ bibliography: paper.bib
 
 This package provides a library of functions as a user-friendly interface 
 for performing time-domain Nuclear Magnetic Resonance (NMR) data processing and visualizations.
-It is aimed towards NMR users who are not necessarily familiar with the julia programming language,
+It is aimed towards NMR users who are not necessarily familiar with the Julia programming language,
 or even programming in general.
 
 Functionality includes importing data from various NMR instrument formats,
-performing phase-correction on raw data, fitting exponential decays using numerical inversions, 
-and visualizing the results in a straightforward manner.
+performing phase-correction on raw data, fitting multi-exponential curves using numerical inversions,
+and interactively visualizing the results in a straightforward manner.
+
+It should be noted that the aforementioned inversions should not be confused with 
+harmonic inversions [@mandel1997] used in NMR spectroscopy. These can be handled by 
+other packages such as "harminv" (https://github.com/NanoComp/harminv), 
+and are not yet implemented in NMRInversions.jl at the time of writing.
 
 # Statement of need
 
 NMR relaxation and diffusion methods are popular in various fields of science 
 and engineering, with applications including studying the properties of 
 porous rocks, catalyst supports and biological tissues, among many others.
+In order to extract useful information from such measurements, numerical 
+inversions are often used to uncover the different relaxation times 
+or diffusion coefficients contributing to the experimental data.
 
-However, many users of such experimental techniques are not fully familliar with
-the technical aspects of the numerical inversioninvolved in the data processing, 
-and sometimes find it challenging to write their own algorithms.
+However, working with the aforementioned inversion methods can be quite daunting towards 
+users who are not very familiar with linear algebra, optimization methods, or programming 
+environments in general.
+This package can cover the needs of users who would like a very simple interface for 
+everyday data processing, without assuming any knowledge on the topic, while retaining 
+the ability to dive into the details and have control over every step of the process 
+if needed, once experience is acquired.
 
-This package can cover the needs of users who would like a very simple interface for everyday data processing, 
-while retaining the ability to dive into the details and have control over every step of the process
-if needed.
-
-The julia programming language [@bezanson2017julia] is an excellent option for such a package, since it 
-provides user-friendly, high level syntax which is similar to MATLAB, whitout any 
+The Julia programming language [@bezanson2017Julia] is an excellent option for such a package, since it 
+provides user-friendly, high level syntax which is similar to MATLAB, without any 
 compromise when it comes to computational performance. 
 It's a thriving ecosystem for scientific computing applications, 
-with several optimization libaries able to tackle the problems arising in NMR applications. 
-
+with several optimization libraries able to tackle the problems arising in NMR applications. 
 Julia's capabilities can enable easy integration with the latest advances from the literature, 
 without the need for writing solvers in low-level languages for performance purposes, 
-in line with julia's philosophy of solving the "two language problem".
-
-Additionally, julia's "multiple dispatch" 
+in line with Julia's philosophy of solving the "two language problem".
+Additionally, Julia's "multiple dispatch" 
 allow us to use the same function names for different 
 operations, depending on the function inputs.
 This enables our package to be more user-friendly, as it 
[…]
