+++ docs/src/types_structs.md
@@ -1,7 +1,4 @@
-
-# Types and Structures
-
-In order to take advantage of julia's multiple dispatch, 
+In order to take advantage of Julia's multiple dispatch, 
 we need to define some structures which can be used as input to our functions.
 
 !!! info
@@ -28,7 +25,8 @@ IRCPMG
 ## Inversion solvers
 These are used to let the invert fucntion know which solver to use.
 They can be used as input to the `invert` function as the 'solver' argument.
-(e.g., `invert(data, solver=brd)` or `invert(data, solver=pdhgm(10,0.1) )`).
+(e.g., `invert(data, solver=brd)` [Song2002](@cite) or 
+`invert(data, solver=pdhgm(10,0.1) )`) [Reci2017](@cite).
 
 ```@docs
 brd
