### 2025-05-05 aris-mav/NMRInversions.jl#59 @cfarm6
Contribution Guidelines
Contribution guidelines are missing per the requirements in the JOSS software submission.
