### 2025-05-05 #7745 @cfarm6
I have reviewed and filed the following issues related to installation instructions and contribution guidelines. Otherwise everything looks good to me.

**Update, I saw the installation issue was just closed. Only remaining piece is a  contribution guidelines section in the docs
[59](https://github.com/aris-mav/NMRInversions.jl/issues/59)
