### 2025-05-05 aris-mav/NMRInversions.jl#59 @aris-mav
This is briefly addressed at the final lines of the [overview](https://aris-mav.github.io/NMRInversions.jl/stable/).
I'd be happy to expand on these points, but I'm not quite sure what to add.
Would you please provide any suggestions or examples?
