### 2025-05-05 727424c
Revert "added git@ to deploydocs"
This reverts commit 360a4e5277490f42e0672e8d67e9135c1f68cd65.
