### 2025-05-05 eeca1cf
added compat in documenter
