+++ docs/Project.toml
@@ -5,3 +5,6 @@ DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
 LiveServer = "16fef848-5104-11e9-1b77-fb7a48bbb589"
 NMRInversions = "55c20db2-0166-4687-95c3-62a9c7afb29b"
+
+[compat]
+Documenter = "1.8.0"
