+++ dev/refs.bib
@@ -77,7 +77,7 @@
     pages = {261-268},
     year = {2002},
     issn = {1090-7807},
-    doi = {https://doi.org/10.1006/jmre.2001.2474},
+    doi = {10.1006/jmre.2001.2474},
     url = {https://www.sciencedirect.com/science/article/pii/S1090780701924747},
     author = {Y.-Q. Song and L. Venkataramanan and M.D. Hürlimann and M. Flaum and P. Frulla and C. Straley},
     abstract = {Spin relaxation is a sensitive probe of molecular structure and dynamics. Correlation of relaxation time constants, such as T1 and T2, conceptually similar to the conventional multidimensional spectroscopy, have been difficult to determine primarily due to the absense of an efficient multidimensional Laplace inversion program. We demonstrate the use of a novel computer algorithm for fast two-dimensional inverse Laplace transformation to obtain T1–T2 correlation functions. The algorithm efficiently performs a least-squares fit on two-dimensional data with a nonnegativity constraint. We use a regularization method to find a balance between the residual fitting errors and the known noise amplitude, thus producing a result that is found to be stable in the presence of noise. This algorithm can be extended to include functional forms other than exponential kernels. We demonstrate the performance of the algorithm at different signal-to-noise ratios and with different T1–T2 spectral characteristics using several brine-saturated rock samples.}
@@ -90,7 +90,7 @@
     pages = {188-198},
     year = {2017},
     issn = {1090-7807},
-    doi = {https://doi.org/10.1016/j.jmr.2017.05.010},
+    doi = {10.1016/j.jmr.2017.05.010},
     url = {https://www.sciencedirect.com/science/article/pii/S1090780717301362},
     author = {A. Reci and A.J. Sederman and L.F. Gladden},
     keywords = { regularization, Inverse problems, 2D NMR correlation experiments, 2D inverse Laplace transformation},
