+++ docs/make.jl
@@ -24,7 +24,8 @@ makedocs(sitename="NMRInversions.jl",
                 "Functions" => "functions.md",
                 "Types and Structures" => "types_structs.md",
                 "Saving data" => "savefiles.md",
-                "References" => "references.md"
+                "References" => "references.md",
+                "Contributing" => "contributing.md"
                ],
          checkdocs=:none,
          format = Documenter.HTML(
