+++ docs/src/contributing.md
@@ -1,0 +1,9 @@
+If you would like to contribute to the package, you may start with one of 
+the [existing issues](https://github.com/aris-mav/NMRInversions.jl/issues).
+Please feel free to create a [pull request](https://docs.github.com/en/pull-requests)
+if you'd like to implement a feature. Stanard [Julia style]
+(https://docs.julialang.org/en/v1/manual/style-guide/) 
+is recommended, and writing tests for any new (or existing) features would be nice.
+
+If you're unsure about anything or need further guidance, feel free to start a
+[discussion](https://github.com/aris-mav/NMRInversions.jl/discussions).
