### 2025-05-05 aris-mav/NMRInversions.jl#59 @aris-mav
Just turned it into a different section (thought it would be more visible this way) and added a few more points.
Please let me know what you think.
