### 2025-05-05 aris-mav/NMRInversions.jl#59 @cfarm6
That looks better. Thanks. I'll mark this as closed.
