### 2025-05-05 8ec0c47
remove contributing section from overview
