### 2025-05-07 a74ba26
added yscale option on 1D plots
