+++ paper/paper.bib
@@ -63,4 +63,5 @@
     pages = {6756-6769},
     year = {1997},
     month = {11},
+    doi = {10.1063/1.475324}
 }
