+++ versions.js
@@ -1,7 +1,8 @@
 var DOC_VERSIONS = [
   "stable",
+  "v1.0",
   "v0.9",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.9.7";
+var DOCUMENTER_NEWEST = "v1.0.0";
 var DOCUMENTER_STABLE = "stable";
