### 2025-05-16 #7745 @aris-mav
Thanks @zhubonan, I confirm that I went through the above.
Release is 1.0.0.
Here is the [zenodo doi](https://doi.org/10.5281/zenodo.15433641).
