+++ paper/paper.bib
@@ -10,6 +10,14 @@
     doi={10.1137/141000671}
 }
 
+@misc{harminvcite,
+    author = {Steven G. Johnson},
+    title = {harminv},
+    url = {https://github.com/NanoComp/harminv},
+    date = {2023},
+    version = {1.4.2}
+}
+
 @article{Fordham2017,
     author = {Fordham, Edmund and Venkataramanan, Lalitha and Mitchell, Jonathan and Valori, Andrea},
     year = {2017},
